# lex name and identifier