```console
$ benchstat a.txt b.txt                             
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/parse
cpu: Apple M2 Max
                         │    a.txt    │               b.txt                │
                         │   sec/op    │   sec/op     vs base               │
ComplexMessage_String-12   4.169µ ± 1%   4.175µ ± 1%       ~ (p=0.698 n=10)
Lex-12                     2.006µ ± 0%   1.931µ ± 1%  -3.74% (p=0.000 n=10)
Parse-12                   3.833µ ± 0%   3.737µ ± 1%  -2.53% (p=0.002 n=10)
geomean                    3.177µ        3.111µ       -2.06%

                         │    a.txt     │                 b.txt                 │
                         │     B/op     │     B/op      vs base                 │
ComplexMessage_String-12   27.16Ki ± 0%   27.16Ki ± 0%       ~ (p=1.000 n=10) ¹
Lex-12                       856.0 ± 0%     856.0 ± 0%       ~ (p=1.000 n=10) ¹
Parse-12                   7.438Ki ± 0%   7.438Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                    5.528Ki        5.528Ki       +0.00%
¹ all samples are equal

                         │   a.txt    │                b.txt                │
                         │ allocs/op  │ allocs/op   vs base                 │
ComplexMessage_String-12   61.00 ± 0%   61.00 ± 0%       ~ (p=1.000 n=10) ¹
Lex-12                     43.00 ± 0%   43.00 ± 0%       ~ (p=1.000 n=10) ¹
Parse-12                   77.00 ± 0%   77.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    58.67        58.67       +0.00%
¹ all samples are equal
```